Fix uint8_t overflow in DisplayWrapper::getStringWidth() callers
DisplayWrapper::getStringWidth() is declared uint8_t but returns
LovyanGFX's int32_t textWidth(), so it wraps modulo 256. On the TFT the
wrapper maps DialogInput_* to IntelOneMono 12pt/15pt - 15 px and 18 px
per character, both strictly monospace - so the wrap starts at 18
characters in the status-line font and at 15 in the scroll-area font,
well inside what a 320 px line holds. "Continue with paddle " (the TFT
generator/echo-trainer pause message) measures 315 px and reported 59.

Where the value only feeds a fillRect() behind text the damage is
masked: LovyanGFX's own glyph background boxes fill exactly the same
rectangle when fore != back colour, and SCROLL_TOP / LINE_HEIGHT are
fontHeight(), which is the same metrics.height those boxes use. Where it
feeds column arithmetic it was live:

  - printToScroll_internal: scrollScreenPos += width / C_WIDTH. Chunks
    are capped at printToScroll_bufMax = 15 on the TFT, so 15 x 18 =
    270 px reported 14, and 14/18 advanced the column by 0 - the next
    chunk overprinted the previous one. Reached by any printToScroll()
    text of 15+ characters: the QSO Bot's "[QSO complete (no RST)]" and
    "(click to exit)\n", English Words at unlimited length (up to 17),
    file-player tokens (up to 127).
  - refreshScrollLine: charsPrinted = printOnScroll(...) / C_WIDTH, the
    same failure when a styled line is redrawn.

Fixed caller-side rather than in the library, which would mean a version
bump - and would not have been enough on its own, since printOnScroll()
and printOnStatusLine() assigned into their own uint8_t locals. New
file-local stringWidth() helper measures through getLGFX()->textWidth()
on the TFT and forwards to display.getStringWidth() on the OLED, which
has no such defect. printOnScroll()'s return type widened to uint16_t
along with the locals.

setTextAlignment() - the other silent stub in that library - needs no
change: both in-tree callers pass TEXT_ALIGN_LEFT, which is what the
empty stub already does. Both traps recorded in devdocs/HANDOFF.md.

Builds clean for pocketwroom, heltec_wifi_lora_32_V2 and
pocketwroom-accessibility. No new UI text, so nothing owed to the
accessibility voice tables (extractor re-run: empty diff).

+++ Software/src/Version 6 and newer/MorseOutput.h
@@ -54,7 +54,10 @@ namespace MorseOutput
   void clearBuffer();
   void refreshScrollArea(int relPos);
   void refreshScrollLine(int bufferLine, int displayLine);
-  uint8_t printOnScroll(uint8_t line, FONT_ATTRIB how, uint8_t xpos, const String& mystring, boolean small = false);
+  /// returns the width of the printed string in pixels - uint16_t, not uint8_t:
+  /// a full scroll line is 306 px on the TFT and callers divide the result by the
+  /// character width to get a column count (see refreshScrollLine).
+  uint16_t printOnScroll(uint8_t line, FONT_ATTRIB how, uint8_t xpos, const String& mystring, boolean small = false);
   void printToScroll(FONT_ATTRIB style, const String& text, boolean autoflush, boolean scroll);
   void printToScroll_internal(FONT_ATTRIB style, const String& text, boolean scroll);
   boolean wordNeedsWrap(uint16_t wordLen);
